+++ src/optimize/CMakeLists.txt
@@ -1,0 +1,12 @@
+# Add sources
+target_sources(${PROJECT_NAME}
+               PRIVATE
+               ${CMAKE_CURRENT_SOURCE_DIR}/Optimizer.cc ${CMAKE_CURRENT_SOURCE_DIR}/Optimizer.h
+               ${CMAKE_CURRENT_SOURCE_DIR}/OptimizerWithoutMiddlePoint.cc ${CMAKE_CURRENT_SOURCE_DIR}/OptimizerWithoutMiddlePoint.h
+               ${CMAKE_CURRENT_SOURCE_DIR}/OptimizerDistanceOnly.cc ${CMAKE_CURRENT_SOURCE_DIR}/OptimizerDistanceOnly.h
+               )
+
+# Install headers
+file(GLOB HEADERS "${CMAKE_CURRENT_SOURCE_DIR}/*.h")
+install(FILES ${HEADERS}
+        DESTINATION ${STBR_INCLUDE_INSTALL_DIR}/optimize)
