+++ src/optimize/Optimizer.cc
