+++ src/optimize/OptimizerWithoutMiddlePoint.cc
