+++ src/optimize/OptimizerWithoutMiddlePoint.h
